requirements.txt
pillow
pytesseract
ffmpeg-python
numpy
nltk